Accept ext_if and int_if as args to bootstrap script
We shouldn't be using `eth0` and `eth1` as hard coded values. Odin should tell us which ones the provider expects, instead, and we should use those accordingly. This is already happening for int_if, to be fair, but ext_if is left entirely unaccounted for. This commit adds support for ext_if, and also ensures both interfaces are used correctly in all the configs that are affected by them (including the NAT setup in the iptables config).